refactor(onboard): raise code health in the Desktop driver
CodeScene failed the PR on three new files: live.go at 4.87 (Low Cohesion,
Complex Method), oracle.go at 6.79 (Bumpy Road Ahead, Complex Method) and
oracle_test.go at 7.47.

live.go was one 1159-line file holding six unrelated concerns. It is now six
files in the same package, split by what each one is responsible for. This is
a pure move — no signature, behaviour or export changes:

  live.go           runtime construction, preflight, composer control
  live_session.go   registry / transcript identity and its recovery
  live_irrlicht.go  what the daemon reports, and the recording it writes
  live_process.go   the process census and the bounded exit wait
  live_evidence.go  capture, cross-check, and staging writes
  live_archive.go   archiving, and the guards around it

Three flagged methods are decomposed:

- validateEvidenceIdentity split into validateRegistryTranscriptLink,
  validateIrrlichtLink and validateEnvironmentLink. It is the join every
  Desktop recording rests on, and each link now fails with its own message
  instead of one shared sentence for five different breaks.
- readManagedFileManifest's row validation extracted to
  parseManagedFileRow, which returns keep=false only for a row the oracle
  does not model and an error for every other malformed shape.
- declaredManagedPaths' inner resolver loop extracted to
  resolveWrittenPaths.

TestManagedFileOracleRunsRealClaudeCodeApplyClosuresInShadowHome was one
80-line case asserting three separate properties. It is now three named
cases over a shared oracleFixture, so a failure says which property broke.

No test was weakened: the same assertions run, and the suite passes with
-race.

Refs #1887
